Separate object_store tests from regular tests
