+++ rust/src/security/sasl.rs
@@ -250,7 +250,11 @@ impl SaslReader {
                 ));
             }
             RpcStatusProto::Fatal => {
-                warn!("RPC fatal error: {:?}", rpc_response.error_msg);
+                warn!(
+                    "RPC fatal error: {}: {}",
+                    rpc_response.exception_class_name(),
+                    rpc_response.error_msg()
+                );
                 return Err(HdfsError::FatalRPCError(
                     rpc_response.exception_class_name().to_string(),
                     rpc_response.error_msg().to_string(),
